Code Review: refactor RelayMessageMixin location; add comments
