+++ cosinnus/views/mixins/tagged.py
@@ -521,4 +521,16 @@ class RecordLastVisitedMixin(object):
                 'was trying to log a LastVisit, but no `self.object` could be found!')
         if self.request.user.is_authenticated and getattr(self, 'object', None) is not None:
             self.object.mark_visited(self.request.user)
-    
+
+
+class RelayMessageMixin:
+    """ Utility function mixin for a BaseTaggableObject model relaying messages to Rocket Chat. """
+
+    def get_message_emote(self):
+        return settings.COSINNUS_ROCKET_NEWS_BOT_EMOTE
+
+    def get_message_title(self):
+        raise NotImplementedError
+
+    def get_message_text(self):
+        raise NotImplementedError
